+++ system/src/Grav/Common/Media/Interfaces/MediaCollectionInterface.php
@@ -4,6 +4,6 @@ namespace Grav\Common\Media\Interfaces;
 /**
  * Class implements media collection interface.
  */
-interface MediaCollectionInterface
+interface MediaCollectionInterface extends \Grav\Framework\Media\Interfaces\MediaCollectionInterface
 {
 }