+++ system/src/Grav/Common/Media/Interfaces/MediaObjectInterface.php
@@ -4,6 +4,6 @@ namespace Grav\Common\Media\Interfaces;
 /**
  * Class implements media object interface.
  */
-interface MediaObjectInterface
+interface MediaObjectInterface extends \Grav\Framework\Media\Interfaces\MediaObjectInterface
 {
 }