+++ system/src/Grav/Framework/Media/Interfaces/MediaCollectionInterface.php
@@ -1,0 +1,9 @@
+<?php
+namespace Grav\Framework\Media\Interfaces;
+
+/**
+ * Class implements media collection interface.
+ */
+interface MediaCollectionInterface
+{
+}