+++ system/src/Grav/Framework/Media/Interfaces/MediaObjectInterface.php
@@ -1,0 +1,9 @@
+<?php
+namespace Grav\Framework\Media\Interfaces;
+
+/**
+ * Class implements media object interface.
+ */
+interface MediaObjectInterface
+{
+}